Fix issue with running reports and old version of guava being version used
If using Maven, and only define a dependency on serenity:serenity-jbehave, Maven
appears to resolve guava to the old version contained in reflections (0.11.0).

This version of Guava does not contain many of the methods required to complete reporting.

Adding this exclusion has all the tests pass, AND my FT suite completes successfully pointing at a
local SNAPSHOT version of this project.